refactor: change the match statement to if/else to account for better condition checking
Changed the "add", "edit", and "complete" features to allow the user to input the item or
number after the menu selection on the same line. This cuts down on extra prompts and
increases ease of use. The with statement cannot check expressions in the case line, that's
why it had to change to if/else.