fix: breaklines missing from adding items and bug in searching strings
When refactoring the "add" feature, the breaklines were inadverdantly
left off at the end of each item. This has been fixed.

When searching for the menu selection in the user input, if the user
supplied a string containing multiple menu options by accident, the program
may match the wrong selection. This was fixed by replacing the "in"
operator with the "startswith()" string function.